+++ tests/integration/conftest.py
@@ -41,4 +41,4 @@ def finetuner_mocker(mocker):
 
     finetuner.login()
 
-    return finetuner.ft
+    yield finetuner.ft