+++ CHANGELOG.md
@@ -57,6 +57,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Enable saving models for clip without overwriting. ([#432](https://github.com/jina-ai/finetuner/pull/432))
 
+- Fix flaky integration test. ([#436](https://github.com/jina-ai/finetuner/pull/436))
+
 ## [0.1.0] - 2022-05-23
 
 ### Added